Moving compute_sh_coefficients from reconst.dwi to reconst.sh, thus removing reconst.dwi now empty.
