--- scilpy/reconst/tests/test_dwi.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-
-def test_compute_sh_coefficients():
-    # toDO
-    pass
