+++ scilpy/reconst/tests/test_sh.py
@@ -1,6 +1,11 @@
 # -*- coding: utf-8 -*-
 
 
+def test_compute_sh_coefficients():
+    # toDO
+    pass
+
+
 def test_compute_rish():
     # toDO
     pass
